AbsoluteBlockNumberFormControl component
